requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.31.0
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.26.0
joblib==1.3.2
firebase-admin==7.2.0
httpx==0.28.1
